backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
PyMySQL==1.1.2
python-dotenv==1.0.1
pydantic==2.8.2
openpyxl==3.1.5
cryptography
